docs(cli): draft user experience for enrollment
